food-insight-app/requirements.txt
streamlit
requests
python-dotenv
pandas
pydantic
Pillow

